requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.9.0
streamlit>=1.34.0
scikit-learn>=1.3.0
pgmpy>=0.1.24
matplotlib>=3.8.0
pandas>=2.0.0
joblib>=1.3.0
datasets>=2.19.0
pyarrow>=15.0.0
